Move ComparisonFile types to tests/types folder
