+++ packages/core/comparison/lib/checkPhase.ts
@@ -4,7 +4,7 @@ import type {
   Phase1Comparison,
   Phase2Comparison,
   Phase3Comparison
-} from "../types/ComparisonFile"
+} from "../../tests/types/ComparisonFile"
 
 export const isPhase1 = (comparison: Comparison): comparison is OldPhase1Comparison | Phase1Comparison =>
   !("phase" in comparison) || comparison.phase === 1
