+++ packages/core/comparison/lib/comparePhase2.test.ts
@@ -1,4 +1,4 @@
-import type { Phase2Comparison } from "../types/ComparisonFile"
+import type { Phase2Comparison } from "../../tests/types/ComparisonFile"
 
 import comparePhase2 from "./comparePhase2"
 import processTestFile from "./processTestFile"
