+++ packages/core/comparison/lib/processTestFile.ts
@@ -3,7 +3,7 @@ import fs from "fs"
 import orderBy from "lodash.orderby"
 import path from "path"
 
-import type { Comparison } from "../types/ComparisonFile"
+import type { Comparison } from "../../tests/types/ComparisonFile"
 
 export const parseComparisonFile = (contents: string, file?: string): Comparison => {
   let parsed: Comparison
